pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "mwparserfromhell>=0.6",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "responses>=0.25",
]

[project.scripts]
wiki-pipeline = "wiki_pipeline.pipeline:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=wiki_pipeline --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["src"]
